Move location helper to a library
Update user resource to use groovy script location helper from library.
Update repo resource to use the same library and move its Groovy scripts into
their own files.
