+++ files/default/delete_repo.groovy
@@ -1,0 +1,6 @@
+def repo = repository.repositoryManager.get(args)
+if (repo == null) {
+    return false
+}
+repository.repositoryManager.delete(args)
+true
